+++ rust/message-backup/src/backup/recipient.rs
@@ -236,7 +236,7 @@ pub enum Destination<R> {
     Group(GroupData),
     #[serde(bound(serialize = "DistributionListItem<R>: serde::Serialize"))]
     DistributionList(DistributionListItem<R>),
-    Self_,
+    Self_(SelfData),
     ReleaseNotes,
     CallLink(CallLink),
 }
@@ -300,6 +300,8 @@ pub struct ContactData {
     pub system_given_name: String,
     pub system_family_name: String,
     pub system_nickname: String,
+    #[serde(serialize_with = "serialize::optional_enum_as_string")]
+    pub avatar_color: Option<proto::AvatarColor>,
 }
 
 #[derive(Clone, Debug, serde::Serialize)]
@@ -307,6 +309,13 @@ pub struct ContactData {
 pub struct ContactName {
     pub given_name: String,
     pub family_name: String,
+}
+
+#[derive(Clone, Debug, serde::Serialize)]
+#[cfg_attr(test, derive(PartialEq))]
+pub struct SelfData {
+    #[serde(serialize_with = "serialize::optional_enum_as_string")]
+    pub avatar_color: Option<proto::AvatarColor>,
 }
 
 #[derive(Clone, Debug, serde::Serialize)]
@@ -378,7 +387,7 @@ impl<R> From<Destination<R>> for MinimalRecipientData {
                     distribution_id, ..
                 },
             ) => Self::DistributionList { distribution_id },
-            Destination::Self_ => Self::Self_,
+            Destination::Self_(_) => Self::Self_,
             Destination::ReleaseNotes => Self::ReleaseNotes,
             Destination::CallLink(CallLink { root_key, .. }) => Self::CallLink { root_key },
         }
@@ -431,7 +440,7 @@ impl<R> AsRef<DestinationKind> for Destination<R> {
             Destination::Contact(_) => &DestinationKind::Contact,
             Destination::Group(_) => &DestinationKind::Group,
             Destination::DistributionList(_) => &DestinationKind::DistributionList,
-            Destination::Self_ => &DestinationKind::Self_,
+            Destination::Self_(_) => &DestinationKind::Self_,
             Destination::ReleaseNotes => &DestinationKind::ReleaseNotes,
             Destination::CallLink(_) => &DestinationKind::CallLink,
         }
@@ -459,7 +468,14 @@ impl<R: Clone, C: LookupPair<RecipientId, MinimalRecipientData, R> + ReportUnusu
             RecipientDestination::DistributionList(list) => {
                 Destination::DistributionList(list.try_into_with(context)?)
             }
-            RecipientDestination::Self_(proto::Self_ { special_fields: _ }) => Destination::Self_,
+            RecipientDestination::Self_(proto::Self_ {
+                avatarColor,
+                special_fields: _,
+            }) => {
+                // The color is allowed to be unset.
+                let avatar_color = avatarColor.map(|v| v.enum_value_or_default());
+                Destination::Self_(SelfData { avatar_color })
+            }
             RecipientDestination::ReleaseNotes(proto::ReleaseNotes { special_fields: _ }) => {
                 Destination::ReleaseNotes
             }
@@ -493,6 +509,7 @@ impl<C: ReportUnusualTimestamp> TryFromWith<proto::Contact, C> for ContactData {
             systemGivenName,
             systemFamilyName,
             systemNickname,
+            avatarColor,
             special_fields: _,
         } = value;
 
@@ -595,6 +612,9 @@ impl<C: ReportUnusualTimestamp> TryFromWith<proto::Contact, C> for ContactData {
             )
             .transpose()?;
 
+        // The color is allowed to be unset.
+        let avatar_color = avatarColor.map(|v| v.enum_value_or_default());
+
         Ok(Self {
             aci,
             pni,
@@ -615,6 +635,7 @@ impl<C: ReportUnusualTimestamp> TryFromWith<proto::Contact, C> for ContactData {
             system_given_name: systemGivenName,
             system_family_name: systemFamilyName,
             system_nickname: systemNickname,
+            avatar_color,
         })
     }
 }
@@ -859,6 +880,7 @@ mod test {
                 system_family_name: "FamilySystemName".to_owned(),
                 system_nickname: "SystemNickName".to_owned(),
                 note: "nb".into(),
+                avatar_color: None,
             }
         }
     }
@@ -882,7 +904,7 @@ mod test {
 
         assert_eq!(
             Destination::try_from_with(recipient, &TestContext::default()),
-            Ok(Destination::Self_)
+            Ok(Destination::Self_(SelfData { avatar_color: None }))
         )
     }